add analysis executor and controller to app_common
